Update rigid dynamic flags

+++ physx/src/rigid_dynamic.rs
@@ -46,12 +46,12 @@ impl PxFlags for RigidDynamicLockFlags {
 
     fn into_px(self) -> Self::Target {
         PxRigidDynamicLockFlags {
-            mBits: self.bits() as u16,
+            mBits: self.bits(),
         }
     }
 
     fn from_px(flags: Self::Target) -> Self {
-        unsafe { BitFlags::from_bits_unchecked(flags.mBits as u8) }
+        unsafe { BitFlags::from_bits_unchecked(flags.mBits) }
     }
 }
